copy updated files from gitee to github

+++ built-in/PyTorch/Official/cv/image_object_detection/RetinaNet_for_PyTorch/mmdet/core/bbox/samplers/sampling_result.py
@@ -45,7 +45,7 @@ class SamplingResult(util_mixins.NiceRepr):
         self.pos_is_gt = gt_flags * pos_inds
 
         self.num_gts = gt_bboxes.shape[0]
-        self.pos_assigned_gt_inds = (assign_result.gt_inds - 1) * pos_inds
+        self.pos_assigned_gt_inds = (assign_result.gt_inds.int() - 1) * pos_inds
 
         if gt_bboxes.numel() == 0:
             # hack for index error case
@@ -58,7 +58,7 @@ class SamplingResult(util_mixins.NiceRepr):
             self.pos_gt_bboxes = torch.index_select(gt_bboxes, 0, self.pos_assigned_gt_inds)
 
         if assign_result.labels is not None:
-            self.pos_gt_labels = assign_result.labels * pos_inds
+            self.pos_gt_labels = (assign_result.labels.int() * pos_inds).long()
         else:
             self.pos_gt_labels = None
